refactor: 패키지 리팩토링
- articleLike > baseLike
