+++ src/main/java/com/mobble/mobbleserver/domain/like/articleLike/service/ArticleLikeQueryServiceImpl.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.domain.like.articleLike.service;
 
 import com.mobble.mobbleserver.domain.article.entity.Article;
 import com.mobble.mobbleserver.domain.article.validator.ArticleValidator;
-import com.mobble.mobbleserver.domain.like.articleLike.dto.response.LikeMemberListResponseDto;
+import com.mobble.mobbleserver.domain.like.baseLike.dto.response.LikeMemberListResponseDto;
 import com.mobble.mobbleserver.domain.like.articleLike.entity.ArticleLike;
 import com.mobble.mobbleserver.domain.like.articleLike.repository.ArticleLikeRepository;
 import com.mobble.mobbleserver.domain.like.baseLike.entity.LikeType;
