+++ src/main/java/com/mobble/mobbleserver/domain/like/baseLike/service/LikeDispatcherService.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.like.baseLike.service;
 
-import com.mobble.mobbleserver.domain.like.articleLike.dto.response.LikeMemberListResponseDto;
+import com.mobble.mobbleserver.domain.like.baseLike.dto.response.LikeMemberListResponseDto;
 import com.mobble.mobbleserver.domain.like.baseLike.dto.response.LikeToggleResponseDto;
 import com.mobble.mobbleserver.domain.like.baseLike.entity.LikeType;
 import com.mobble.mobbleserver.domain.member.entity.Member;
