+++ src/main/java/com/mobble/mobbleserver/domain/like/baseLike/service/LikeQueryService.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.like.baseLike.service;
 
-import com.mobble.mobbleserver.domain.like.articleLike.dto.response.LikeMemberListResponseDto;
+import com.mobble.mobbleserver.domain.like.baseLike.dto.response.LikeMemberListResponseDto;
 import com.mobble.mobbleserver.domain.like.baseLike.entity.LikeType;
 
 public interface LikeQueryService {
